externalized file-loaded-hook
