+++ frontend/src/components/Logo.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useRef, useEffect } from "react";
-import { useSettings, useLogo } from "../hooks";
+import { useSettings, useLogo, useFileLoaded } from "../hooks";
 import defaultLogo from "../logo.svg";
 import { useTranslation } from "react-i18next";
 
@@ -9,20 +9,7 @@ function Logo() {
   const { logo, loadingFile } = useLogo(settings.customLogoS3Key);
   const [toHide, setToHide] = useState<boolean>(true);
 
-  // firstUpdate stops useEffect from executing after the first render.
-  const firstUpdate = useRef(true);
-  const secondUpdate = useRef(true);
-  useEffect(() => {
-    if (secondUpdate.current) {
-      if (firstUpdate.current) {
-        firstUpdate.current = false;
-        return;
-      }
-      secondUpdate.current = false;
-      return;
-    }
-    setToHide(false);
-  }, [loadingFile]);
+  useFileLoaded(setToHide, loadingFile);
 
   return (
     <>
